display: NightDisplay: Use chromatic adaptation when possible
Night Light is currently implemented with a simple RGB tint (similar to
the "wrong von Kries transform" but in linear sRGB instead of XYZ),
which distorts some colors in addition to changing the neutral CCT.
Switch to using full-blown chromatic adaptation (CAT16 + von Kries
transform) instead to improve color quality when Night Light is active.

Because accurate chromatic adaptation is only possible in linear color
spaces, the old logic still needs to be used as a fallback in cases
where color transformations can only be applied in the native (device)
color space. This uses the existing CCT->RGB tint coefficients for the
native color space, which is the only option because we can't make
assumptions about the native color space (even if it's sRGB, we can't
accommodate it with only matrices).

Change-Id: I762610ed3c2cc417307485fd0a6a6d499b00c7d5
Signed-off-by: rezaadi0105 <rezaadipangestu5@gmail.com>